+++ Java Game/src/main/GameObject.java
@@ -5,11 +5,11 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velX, velY; //controls speed of x and y directions
+	protected float velX, velY; //controls speed of x and y directions
 	
-	public GameObject(int x, int y, ID id){
+	public GameObject(float x, float y, ID id){
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -26,10 +26,10 @@ public abstract class GameObject {
 	public void setY(int y){
 		this.y=y;
 	}
-	public int getX(){
+	public float getX(){
 		return x;
 	}
-	public int getY(){
+	public float getY(){
 		return y;
 	}
 	public void setID(ID id){
@@ -44,10 +44,10 @@ public abstract class GameObject {
 	public void setVelY(int velY){
 		this.velY=velY;
 	}
-	public int getVelX(){
+	public float getVelX(){
 		return velX;
 	}
-	public int getVelY(){
+	public float getVelY(){
 		return velY;
 	}
 }
